Move box_filter into a separate file
